split to components
